moved context and provider to seperate @/contexts folder
